rename lib files
