+++ include/leo/leo_internal.h
@@ -85,6 +85,10 @@ typedef struct {
 #define LEO_COUNTRY_USA		0x2263EE56
 #define LEO_COUNTRY_NONE	0x00000000
 
+#define LEO_CMD_TYPE_0 0
+#define LEO_CMD_TYPE_1 1
+#define LEO_CMD_TYPE_2 2
+
 #define LEO_STATUS_PRESENCE_MASK 0xFFFF
 
 #define LEO_STATUS_DATA_REQUEST 0x40000000
