+++ src/game/10490.c
@@ -16,7 +16,7 @@ void func_80076498(u32 arg0, u32 arg1, u32 arg2) {
     D_800DCCA8.dramAddr = osPhysicalToVirtual(arg1);
     D_800DCCA8.devAddr = arg0;
     D_800DCCA8.size = arg2;
-    D_800DCCDC->transferInfo.cmdType = 2;
+    D_800DCCDC->transferInfo.cmdType = LEO_CMD_TYPE_2;
     osEPiStartDma(D_800DCCDC, &D_800DCCA8, 0);
     MQ_WAIT_FOR_MESG(&D_800DCA68, sp1C);
 }
@@ -29,7 +29,7 @@ void func_80076528(u32 arg0, u32 arg1, u32 arg2, void* arg3, s32 arg4) {
     D_800DCCA8.dramAddr = osPhysicalToVirtual(arg1);
     D_800DCCA8.devAddr = arg0;
     D_800DCCA8.size = arg2;
-    D_800DCCDC->transferInfo.cmdType = 2;
+    D_800DCCDC->transferInfo.cmdType = LEO_CMD_TYPE_2;
     osEPiStartDma(D_800DCCDC, &D_800DCCA8, 0);
     bzero(arg3, arg4);
     MQ_WAIT_FOR_MESG(&D_800DCA68, sp24);
