+++ src/game/73F0.c
@@ -602,7 +602,7 @@ void func_80073E28(void* arg0, void* arg1, size_t arg2) {
     D_800DCCA8.dramAddr = osPhysicalToVirtual(arg1);
     D_800DCCA8.devAddr = (uintptr_t) arg0;
     D_800DCCA8.size = arg2;
-    D_800DCCDC->transferInfo.cmdType = 2;
+    D_800DCCDC->transferInfo.cmdType = LEO_CMD_TYPE_2;
     osEPiStartDma(D_800DCCDC, &D_800DCCA8, 0);
     MQ_WAIT_FOR_MESG(&D_800DCA68, NULL);
 }
@@ -619,7 +619,7 @@ void func_80073ED0(void* arg0, void* arg1, size_t arg2) {
     D_800DCCA8.dramAddr = osPhysicalToVirtual(arg1);
     D_800DCCA8.devAddr = (uintptr_t) arg0;
     D_800DCCA8.size = arg2;
-    D_800DCCDC->transferInfo.cmdType = 2;
+    D_800DCCDC->transferInfo.cmdType = LEO_CMD_TYPE_2;
     osEPiStartDma(D_800DCCDC, &D_800DCCA8, 0);
     MQ_WAIT_FOR_MESG(&D_800DCA68, sp20);
 }
