+++ src/leo/lib/bytetolba.c
