+++ src/leo/lib/leoinquiry.c
