+++ src/leo/lib/leoreset.c
