+++ src/leo/lib/leorezero.c
