+++ src/leo/lib/leoseek.c
