+++ src/leo/lib/leotimer.c
@@ -5,80 +5,6 @@
 u8 __locReadTimer(__LOCTime* time);
 u8 __locSetTimer(__LOCTime* time);
 
-void leoTranslate(void) {
-    u32 lba;
-    u32 calc_bytes;
-    u32 calc_blks;
-    u32 byte_p_blk;
-    u16 zone;
-    u16 vzone;
-    u8 flag; // boolean
-    if (LEOcur_command->data.readwrite.lba >= NUM_LBAS) {
-
-        LEOcur_command->header.sense = LEO_SENSE_LBA_OUT_OF_RANGE;
-        LEOcur_command->header.status = LEO_STATUS_CHECK_CONDITION;
-        return;
-    }
-    lba = LEOcur_command->data.readwrite.lba + 0x18;
-    calc_blks = 0;
-    calc_bytes = 0;
-    flag = vzone = 1;
-
-    if (LEOcur_command->header.control & LEO_CONTROL_TBL) {
-        calc_bytes = LEOcur_command->data.readwrite.xfer_blks;
-
-        while (calc_bytes != 0) {
-            if (flag || (LEOVZONE_TBL[LEOdisk_type][vzone] == lba)) {
-                vzone = leoLba_to_vzone(lba);
-                zone = LEOVZONE_PZONEHD_TBL[LEOdisk_type][vzone];
-                if (zone >= 8) {
-                    zone -= 7;
-                }
-                byte_p_blk = LEOBYTE_TBL2[zone];
-            }
-            if (calc_bytes < byte_p_blk) {
-                calc_bytes = 0;
-            } else {
-                calc_bytes -= byte_p_blk;
-            }
-            calc_blks++;
-            lba++;
-
-            if ((calc_bytes != 0) && (lba >= 0x10DC)) {
-                LEOcur_command->header.sense = LEO_SENSE_LBA_OUT_OF_RANGE;
-                LEOcur_command->header.status = LEO_STATUS_CHECK_CONDITION;
-                return;
-            }
-            flag = false;
-        }
-        LEOcur_command->data.readwrite.buff_ptr = calc_blks;
-    } else {
-        calc_blks = LEOcur_command->data.readwrite.xfer_blks;
-
-        while (calc_blks != 0) {
-            if (flag || (LEOVZONE_TBL[LEOdisk_type][vzone] == lba)) {
-                vzone = leoLba_to_vzone(lba);
-                zone = LEOVZONE_PZONEHD_TBL[LEOdisk_type][vzone];
-                if (zone >= 8) {
-                    zone -= 7;
-                }
-                byte_p_blk = LEOBYTE_TBL2[zone];
-            }
-            calc_bytes += byte_p_blk;
-            calc_blks--;
-            lba++;
-            if ((calc_blks != 0) && (lba >= 0x10DC)) {
-                LEOcur_command->header.sense = LEO_SENSE_LBA_OUT_OF_RANGE;
-                LEOcur_command->header.status = LEO_STATUS_CHECK_CONDITION;
-                return;
-            }
-            flag = false;
-        }
-        LEOcur_command->data.readwrite.buff_ptr = calc_bytes;
-    }
-    LEOcur_command->header.status = LEO_STATUS_GOOD;
-}
-
 void leoReadTimer(void) {
     u8* rdparam;
     u8 data[4];
@@ -97,10 +23,10 @@ void leoReadTimer(void) {
         LEOcur_command->header.status = LEO_STATUS_CHECK_CONDITION;
         return;
     }
-    if ((u8)LEOcur_command->data.time.yearlo >= 0x96U) {
-        LEOcur_command->data.time.yearhi = 0x19;
+    if ((u32)LEOcur_command->data.time.yearlo >= 150) {
+        LEOcur_command->data.time.yearhi = 25;
     } else {
-        LEOcur_command->data.time.yearhi = 0x20;
+        LEOcur_command->data.time.yearhi = 32;
     }
     LEOcur_command->header.status = 0;
 }
