+++ src/leo/lib/leoutil.c
