+++ src/leo/lib/readdiskid.c
