+++ src/leo/lib/seek.c
