+++ src/leo/lib/spdlmotor.c
